* next: (953 commits)
  Add paragraph on setting staggered location for Laplacian solvers
  Update info on allowed staggering in derivatives
  Fix reference to sec-iterating
  Add MUMPS removal to list of breaking changes
  Typofix
  fix typo
  Add dash as travis job
  Avoid bashism += in configure
  Fix formatting of 'Obtaining BOUT++' section heading
  Remove MUMPS
  prefer PETSc over petsc with pkg-config
  Find PETSc with pkg-conf using configure
  ensure quiet script isn't to verbose on error
  Don't use old-style-cast
  Adjust field compat messages for fmt
  Better debug messages for fields compatibility
  fix requires in options-netcdf test
  improve error messages
  Ensure ffts are OpenMP thread safe
  Improve CPU detection in run_wrapper
  ...

+++ include/bout/fieldgroup.hxx
@@ -153,12 +153,12 @@ public:
 
   /// Return number of fields
   int size() const {
-    return fvec.size();
+    return static_cast<int>(fvec.size());
   }
 
   /// Return number of Field3Ds
   int size_field3d() const {
-    return f3vec.size();
+    return static_cast<int>(f3vec.size());
   }
 
   /// Test whether this group is empty
